requirements.txt
anyio==4.0.0
appnope==0.1.3
asgiref==3.6.0
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
comm==0.1.3
config==0.5.1
debugpy==1.6.7
decorator==5.1.1
Django==4.2.1
django-tinymce4-lite==1.8.0
executing==1.2.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
ipykernel==6.22.0
ipython==8.12.0
jedi==0.18.2
jsmin==3.0.1
jupyter_client==8.2.0
jupyter_core==5.3.0
loader==2017.9.11
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
packaging==23.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.3.0
prompt-toolkit==3.0.38
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.15.1
pyTelegramBotAPI==4.13.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-telegram-bot==20.5
pyzmq==25.0.2
requests==2.31.0
services==0.1.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.2
telebot==0.0.5
telegram==0.0.1
tornado==6.3.1
traitlets==5.9.0
urllib3==2.0.4
wcwidth==0.2.6
